move everything to different features directories
