+++ web/src/features/EthWallet/types/eip6963.d.ts
