+++ web/src/features/EthWallet/types/index.ts
